# evmrs: reuse memory